+++ src/main/java/Duke/DukeCommandLine/DukeTaskInputException.java
@@ -1,4 +1,4 @@
-package DukeCommandLine;
+package Duke.DukeCommandLine;
 
 public class DukeTaskInputException extends Exception {
     public DukeTaskInputException(String message) {
